refactor: mapper로 dto 변환
